format & fix bug
